fix: cargo clippy
