+++ collab-document/src/block_parser/mod.rs
@@ -1,11 +1,9 @@
-pub mod block_parser;
 pub mod document_parser;
 pub mod parsers;
 pub mod registry;
 pub mod text_utils;
 pub mod traits;
 
-pub use block_parser::*;
 pub use document_parser::*;
 pub use parsers::*;
 pub use registry::*;
