+++ collab-document/src/block_parser/parsers/code_block.rs
@@ -29,7 +29,7 @@ impl BlockParser for CodeBlockParser {
         Value::String(s) => Some(s.clone()),
         _ => None,
       })
-      .unwrap_or_else(|| "".to_string());
+      .unwrap_or_default();
 
     let formatted_content = match context.format {
       OutputFormat::Markdown => {
