+++ collab-document/src/block_parser/parsers/subpage.rs
@@ -22,7 +22,7 @@ impl BlockParser for SubpageParser {
         Value::String(s) => Some(s.clone()),
         _ => None,
       })
-      .unwrap_or_else(|| "".to_string());
+      .unwrap_or_default();
 
     let formatted_content = match context.format {
       OutputFormat::Markdown => {
