+++ collab-document/tests/block_parser/todo_list_test.rs
@@ -149,7 +149,6 @@ fn test_todo_list_parser_with_indentation() {
   let block = create_todo_list_block(&mut test, "Indented task".to_string(), Some(true), "");
   let document_data = test.get_document_data();
 
-
   let context = ParseContext::new(&document_data, OutputFormat::Markdown).with_depth(2);
 
   let result = parser.parse(&block, &context).unwrap();
