+++ collab-importer/src/notion/page.rs
@@ -5,7 +5,7 @@ use collab_database::database::{Database, get_row_document_id};
 use collab_database::template::csv::{CSVResource, CSVTemplate};
 use collab_document::blocks::{TextDelta, mention_block_data, mention_block_delta};
 use collab_document::document::Document;
-use collab_document::importer::define::{BlockType, URL_FIELD};
+use collab_document::importer::define::URL_FIELD;
 use collab_document::importer::md_importer::{MDImporter, create_image_block};
 use collab_entity::CollabType;
 use futures::stream::{self, StreamExt};
