+++ collab-importer/tests/notion_test/customer_import_test.rs
@@ -1,5 +1,4 @@
 use crate::util::sync_unzip_asset;
-use collab_document::importer::define::BlockType;
 use collab_importer::notion::NotionImporter;
 
 /// Customer import test 1
